# backend/api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Swyngora Market API
  description: |
    Public market data for Swyngora (v0.1).

    - Candles, 24h volume/ticker, and spot market listing support **binance**, **coinbase**,
      and **bybit** (select with `?exchange=`).
    - Circulating / total / max supply is sourced from the **Binance** public marketing
      symbol list (asset-level, used for all venues' mcap enrichment), refreshed daily.
      Max supply is null when undefined.
    - Watchlist is unauthenticated state keyed by client-supplied `clientId`
      (must be non-empty; not for multi-tenant production without real auth).
      Data is persisted in SQLite on the server (survives restarts). Each list has a monotonic
      `version` for multi-device sync: send `baseVersion` on writes; concurrent adds of different
      symbols auto-merge; delete-vs-update on the same symbol returns 409 with both sides.
    - Price alerts support `one_time` (default) or `repeating` modes for above/below targets.
      One-time alerts become `triggered` after the first fire; repeating alerts stay `active`
      and re-fire only after the price returns to the safe side and crosses again.
    - Clients may register a webhook URL with deliveryMode `immediate` or `hourly_digest`,
      plus optional quiet hours (IANA timeZone + local start/end). Delivery is deferred until
      quiet hours end; pending notifications survive restarts.
...
      orders, and performance), market buy/sell at last price,
...
servers:
...
paths:
...
  /api/v1/portfolio/margin/mode:
    put:
      summary: Set paper margin mode (isolated or cross)
      description: |
        Account-wide margin mode. Cannot change while any open margin position or pending
        margin limit order exists. Default is isolated.
      operationId: setMarginMode
      tags: [Portfolio]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [mode]
              properties:
                clientId: { type: string }
                mode: { type: string, enum: [isolated, cross] }
      responses:
        '200':
          description: Mode updated
        '400':
          $ref: '#/components/responses/Error'
  /api/v1/portfolio/margin/orders:
    post:
      summary: Open paper margin long/short (market or limit)
      description: |
        Uses account marginMode (isolated or cross). Leverage 1–10x. Market opens immediately;
        limit rests until last crosses limitPrice and reserves required margin until fill/cancel.
        Initial margin = qty * price / leverage from available cash.
        Optional stopLoss / takeProfit. Simulated only — not real money.
      operationId: placeMarginOrder
      tags: [Portfolio]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [symbol, side, quantity, leverage]
              properties:
                clientId: { type: string }
                exchange: { type: string, enum: [binance, coinbase, bybit], default: binance }
                symbol: { type: string }
                side: { type: string, enum: [long, short] }
                type: { type: string, enum: [market, limit], default: market }
                quantity: { type: number, exclusiveMinimum: 0 }
                leverage: { type: integer, minimum: 1, maximum: 10 }
                limitPrice: { type: number, exclusiveMinimum: 0, description: Required for type=limit }
                stopLoss: { type: number, exclusiveMinimum: 0 }
                takeProfit: { type: number, exclusiveMinimum: 0 }
      responses:
        '201':
          description: Position opened (market) or limit order resting
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    get:
      summary: List paper margin orders
      operationId: listMarginOrders
      tags: [Portfolio]
      parameters:
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [open, filled, canceled, rejected, all], default: open }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Order list
  /api/v1/portfolio/margin/orders/{id}:
    delete:
      summary: Cancel an open paper margin limit order
      operationId: cancelMarginOrder
      tags: [Portfolio]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      responses:
        '200':
          description: Canceled
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/portfolio/margin/positions:
    get:
      summary: List open paper margin positions
      operationId: listMarginPositions
      tags: [Portfolio]
      parameters:
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      responses:
        '200':
          description: Open margin positions with marks and liquidation price
  /api/v1/portfolio/margin/positions/{id}:
    get:
      summary: Get a paper margin position
      operationId: getMarginPosition
      tags: [Portfolio]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      responses:
        '200':
          description: Position
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarginPosition'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/portfolio/margin/positions/{id}/close:
    post:
      summary: Close all or part of a paper margin position at market
      description: Partial close releases proportional margin and recalculates liquidation.
      operationId: closeMarginPosition
      tags: [Portfolio]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: number
                  description: Omit or 0 for full close
      responses:
        '200':
          description: Updated or closed position plus close trade
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/portfolio/margin/positions/{id}/margin:
    post:
      summary: Add or remove margin from an isolated position
      description: |
        delta > 0 adds cash into the position margin; delta < 0 returns excess to cash.
        Cannot go below initial margin for remaining size. Liquidation price is recalculated.
        Only allowed in isolated mode.
      operationId: adjustMargin
      tags: [Portfolio]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [delta]
              properties:
                delta: { type: number, description: Positive to add, negative to remove }
      responses:
        '200':
          description: Updated position
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/portfolio/margin/positions/{id}/repay:
    post:
      summary: Repay margin debt without closing
      description: |
        Pays interest first, then principal. Amount is in debt units (quote cash for long, base coins for short).
        Short repay spends available cash at mark to cover coin debt.
      operationId: repayMarginDebt
      tags: [Portfolio]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: number, exclusiveMinimum: 0 }
      responses:
        '200':
          description: Updated position and repay trade
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/portfolio/margin/positions/{id}/brackets:
    put:
      summary: Set or clear stop-loss / take-profit on an open margin position
      operationId: setMarginBrackets
      tags: [Portfolio]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stopLoss: { type: number }
                takeProfit: { type: number }
                clearStopLoss: { type: boolean }
                clearTakeProfit: { type: boolean }
      responses:
        '200':
          description: Updated position
        '400':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/v1/portfolio/margin/trades:
    get:
      summary: List paper margin trade history
      operationId: listMarginTrades
      tags: [Portfolio]
      parameters:
        - name: clientId
          in: query
          schema: { type: string }
        - name: X-Client-Id
          in: header
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Margin trades
...
components:
  responses:
    Error:
      description: Error payload
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    MarginPosition:
      type: object
      properties:
        id: { type: string }
        clientId: { type: string }
        exchange: { type: string }
        symbol: { type: string }
        side: { type: string, enum: [long, short] }
        mode: { type: string, enum: [isolated, cross] }
        quantity: { type: number }
        entryPrice: { type: number }
        leverage: { type: integer }
        margin: { type: number }
        debtPrincipal:
          type: number
          description: Main borrowed amount (quote cash for long, base coins for short)
        debtInterest:
          type: number
          description: Accrued interest in the same unit as debtPrincipal
        debtAsset: { type: string, enum: [quote, base] }
        debtNotional:
          type: number
          description: Principal+interest valued in quote (short uses mark)
        lastInterestAt: { type: string, format: date-time }
        liquidationPrice: { type: number }
        stopLoss: { type: number, nullable: true }
        takeProfit: { type: number, nullable: true }
        status: { type: string, enum: [open, closed] }
        markPrice: { type: number }
        unrealizedPnL: { type: number }
        realizedPnL: { type: number }
        closeReason: { type: string }
        openedAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        closedAt: { type: string, format: date-time, nullable: true }
...